requirements.txt
torch
torchvision
scikit-learn
pillow
matplotlib
seaborn
streamlit
numpy